+++ pkgs/google_generative_ai/lib/google_generative_ai.dart
@@ -59,11 +59,21 @@ export 'src/api.dart'
         SafetySetting,
         TaskType;
 export 'src/chat.dart' show ChatSession, StartChatExtension;
-export 'src/content.dart' show Content, DataPart, FilePart, Part, TextPart;
+export 'src/content.dart'
+    show
+        Content,
+        DataPart,
+        FilePart,
+        FunctionCall,
+        FunctionResponse,
+        Part,
+        TextPart;
 export 'src/error.dart'
     show
         GenerativeAIException,
         InvalidApiKey,
         ServerException,
         UnsupportedUserLocation;
+export 'src/function_calling.dart'
+    show FunctionDeclaration, Schema, SchemaType, Tool;
 export 'src/model.dart' show GenerativeModel, RequestOptions;